Compares current version with master branch

+++ Service.py
@@ -4,6 +4,10 @@
 import requests
 import json
 import urllib
+import os
+import re
+ver = "1.0.4"
+
 
 class Service():
     @staticmethod
@@ -53,9 +57,34 @@ class Service():
     @staticmethod
     def getUpdate():
        
-        urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Service.py", "Service.py")
-        print("Downloaded Service.py")
-        urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Episode.py", "Episode.py")
-        print("Downloaded Episode.py")
-        urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Renamer.py", "renamer.py")
-        print("Update complete. Closing.")
+        # Downloads and opens README.md from master branch
+        urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/README.md", "ren_temp.md")
+        file = open('ren_temp.md')
+
+        # Gets the version number from the first line of README
+        master_ver = re.sub('[# Renamer]', '', file.readline())
+
+        print("\nMaster branch is: " + master_ver)
+
+        # Deletes master branch README.md
+        os.remove("ren_temp.md")
+
+        # Ouputs currently running version
+        update = raw_input("Currently on: v" + ver + "\nAre you sure? (y/n): ")
+
+
+        if update.upper() == "Y":
+
+            print("Getting update...")
+            '''
+            urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Service.py", "Service.py")
+            print("Downloaded Service.py")
+             urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Episode.py", "Episode.py")
+            print("Downloaded Episode.py")
+            urllib.urlretrieve ("https://raw.githubusercontent.com/jclishman/Renamer/master/Renamer.py", "renamer.py")
+            print("Update complete. Closing.")
+            '''
+
+        else:
+            print("Update aborted, exiting")
+            exit()
